backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
requests==2.31.0
httpx>=0.26,<0.29
PyYAML==6.0.2
paddlepaddle==2.6.2
paddleocr==2.7.0.3
Pillow==10.2.0
numpy<2.0.0
sentence-transformers==2.7.0
supabase>=2.28,<3
python-dotenv==1.0.1
pytest==8.1.1
celery==5.3.6
redis==5.0.1
